api/asset_manager/src/requirements/requirements.txt
aerich>=0.8.0
fastapi[all]>=0.115.5
python-dotenv>=0.21.0
tortoise-orm[asyncpg]>=0.22.1
uvicorn>=0.31.1
black>=24.10.0
joserfc>=1.0.1
passlib>=1.7.4
pytz>=2024.2
ptpython>=0.25

# Test Suite
httpx>=0.28.1
pytest>=8.3.4
mock>=5.1.0
pytest-mock>=3.14.0
anyio>=4.8.0